feat: replace scrollview with flatlist on RN
