+++ morph/react-native/block-wrap.js
@@ -5,6 +5,7 @@ import wrap from '../react/wrap.js'
 
 export const enter = (node, parent, state) => {
   const name = getBlockName(node, parent, state)
+  debugger
   if (
     name === 'Text' &&
     parent &&
