+++ morph/react-native/block.js
@@ -4,8 +4,8 @@ import * as BlockExplicitChildren from '../react/block-explicit-children.js'
 // import * as BlockGoTo from './block-go-to.js'
 import * as BlockInList from '../react/block-in-list.js'
 import * as BlockName from './block-name.js'
-// import * as BlockList from '../react/block-list.js'
-import * as BlockList from './block-list.js'
+import * as BlockListBasic from './block-list-basic.js'
+import * as BlockListFlat from './block-list-flat.js'
 import * as BlockGroup from '../react/block-group.js'
 import * as BlockProperties from './block-properties.js'
 import * as BlockProxy from '../react/block-proxy.js'
@@ -30,12 +30,14 @@ export const enter = [
   BlockInList.enter,
   BlockAddTestIdProp.enter,
   BlockProperties.enter,
-  BlockList.enter,
   BlockGroup.enter,
+  BlockListBasic.enter,
+  BlockListFlat.enter,
 ]
 
 export const leave = [
-  // BlockList.leave,
+  BlockListBasic.leave,
+  BlockListFlat.leave,
   BlockExplicitChildren.leave,
   BlockName.leave,
   BlockWrap.leave,
