+++ morph/react-native/get-block-name.js
@@ -65,8 +65,8 @@ const getGroupBlockName = (node, state) => {
 
 const getListBlockName = node => {
   const base = hasProp(node, /^overflow/, v => v === 'auto' || v === 'scroll')
-    ? 'ScrollView'
-    : 'FlatList'
+    ? 'FlatList'
+    : 'View'
   return node.maybeAnimated ? `Animated.${base}` : base
 }
