+++ morph/react/block-explicit-children.js
@@ -1,4 +1,5 @@
 export function leave(node, parent, state) {
+  debugger
   if (node.explicitChildren) {
     state.render.push('>')
     state.render.push(node.explicitChildren)
