+++ morph/react/block-group.js
@@ -1,5 +1,10 @@
 export function enter(node, parent, state) {
-  if (node.isGroup && node.children.length > 0) {
+  debugger
+  if (
+    node.isGroup &&
+    node.children.length > 0 &&
+    node.nameFinal !== 'FlatList'
+  ) {
     state.render.push('>')
   }
 }
